refactor: break up customerioclient
