+++ sdk/src/main/java/io/customer/sdk/CustomerIO.kt
@@ -3,12 +3,14 @@ package io.customer.sdk
 import android.app.Activity
 import android.app.Application
 import android.content.pm.PackageManager
-import io.customer.sdk.api.CustomerIOApi
 import io.customer.sdk.data.model.CustomAttributes
 import io.customer.sdk.data.communication.CustomerIOUrlHandler
 import io.customer.sdk.data.model.Region
 import io.customer.sdk.data.request.MetricEvent
 import io.customer.sdk.di.CustomerIOComponent
+import io.customer.sdk.repository.DeviceRepository
+import io.customer.sdk.repository.ProfileRepository
+import io.customer.sdk.repository.TrackRepository
 import io.customer.sdk.util.CioLogLevel
 
 /**
@@ -171,8 +173,14 @@ class CustomerIO internal constructor(
         }
     }
 
-    private val api: CustomerIOApi
-        get() = diGraph.buildApi()
+    private val trackRepository: TrackRepository
+        get() = diGraph.trackRepository
+
+    private val deviceRepository: DeviceRepository
+        get() = diGraph.deviceRepository
+
+    private val profileRepository: ProfileRepository
+        get() = diGraph.profileRepository
 
     override val siteId: String
         get() = diGraph.sdkConfig.siteId
@@ -206,7 +214,7 @@ class CustomerIO internal constructor(
     override fun identify(
         identifier: String,
         attributes: CustomAttributes
-    ) = api.identify(identifier, attributes)
+    ) = profileRepository.identify(identifier, attributes)
 
     /**
      * Track an event
@@ -225,7 +233,7 @@ class CustomerIO internal constructor(
     override fun track(
         name: String,
         attributes: CustomAttributes
-    ) = api.track(name, attributes)
+    ) = trackRepository.track(name, attributes)
 
     /**
      * Track screen
@@ -243,7 +251,7 @@ class CustomerIO internal constructor(
     override fun screen(
         name: String,
         attributes: CustomAttributes
-    ) = api.screen(name, attributes)
+    ) = trackRepository.screen(name, attributes)
 
     /**
      * Track activity screen, `label` added for this activity in `manifest` will be utilized for tracking
@@ -271,19 +279,19 @@ class CustomerIO internal constructor(
      * If no profile has been identified yet, this function will ignore your request.
      */
     override fun clearIdentify() {
-        api.clearIdentify()
+        profileRepository.clearIdentify()
     }
 
     /**
      * Register a new device token with Customer.io, associated with the current active customer. If there
      * is no active customer, this will fail to register the device
      */
-    override fun registerDeviceToken(deviceToken: String) = api.registerDeviceToken(deviceToken, deviceAttributes)
+    override fun registerDeviceToken(deviceToken: String) = deviceRepository.registerDeviceToken(deviceToken, deviceAttributes)
 
     /**
      * Delete the currently registered device token
      */
-    override fun deleteDeviceToken() = api.deleteDeviceToken()
+    override fun deleteDeviceToken() = deviceRepository.deleteDeviceToken()
 
     /**
      * Track a push metric
@@ -292,7 +300,7 @@ class CustomerIO internal constructor(
         deliveryID: String,
         event: MetricEvent,
         deviceToken: String,
-    ) = api.trackMetric(
+    ) = trackRepository.trackMetric(
         deliveryID = deliveryID,
         event = event,
         deviceToken = deviceToken
