+++ sdk/src/sharedTest/java/io/customer/sdk/repository/DeviceRepositoryTest.kt
@@ -1,0 +1,104 @@
+package io.customer.sdk.repository
+
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import io.customer.common_test.BaseTest
+import io.customer.sdk.data.request.Device
+import io.customer.sdk.queue.Queue
+import io.customer.sdk.util.Logger
+import io.customer.sdk.utils.random
+import org.amshove.kluent.shouldBeEqualTo
+import org.junit.Before
+import org.junit.Test
+import org.junit.runner.RunWith
+import org.mockito.kotlin.mock
+import org.mockito.kotlin.verify
+import org.mockito.kotlin.verifyNoInteractions
+
+@RunWith(AndroidJUnit4::class)
+class DeviceRepositoryTest : BaseTest() {
+
+    private val prefRepository: PreferenceRepository
+        get() = di.sharedPreferenceRepository
+    private val backgroundQueueMock: Queue = mock()
+    private val loggerMock: Logger = mock()
+
+    private lateinit var repository: DeviceRepository
+
+    @Before
+    override fun setup() {
+        super.setup()
+
+        repository = DeviceRepositoryImpl(
+            config = cioConfig,
+            deviceStore = deviceStore,
+            preferenceRepository = prefRepository,
+            backgroundQueue = backgroundQueueMock,
+            dateUtil = dateUtilStub,
+            logger = loggerMock
+        )
+    }
+
+    // registerDeviceToken
+
+    @Test
+    fun registerDeviceToken_givenNoIdentifiedProfile_expectDoNotAddTaskToBackgroundQueue_expectSaveToken() {
+        val givenDeviceToken = String.random
+
+        repository.registerDeviceToken(givenDeviceToken, emptyMap())
+
+        verifyNoInteractions(backgroundQueueMock)
+        prefRepository.getDeviceToken() shouldBeEqualTo givenDeviceToken
+    }
+
+    @Test
+    fun registerDeviceToken_givenIdentifiedProfile_expectAddTaskToQueue_expectSaveToken() {
+        val givenIdentifier = String.random
+        val givenDeviceToken = String.random
+        val givenAttributes = mapOf("name" to String.random)
+        prefRepository.saveIdentifier(givenIdentifier)
+
+        repository.registerDeviceToken(givenDeviceToken, givenAttributes)
+
+        verify(backgroundQueueMock).queueRegisterDevice(
+            givenIdentifier,
+            Device(
+                token = givenDeviceToken,
+                lastUsed = dateUtilStub.givenDate,
+                attributes = deviceStore.buildDeviceAttributes() + givenAttributes
+            )
+        )
+        prefRepository.getDeviceToken() shouldBeEqualTo givenDeviceToken
+    }
+
+    // deleteDeviceToken
+
+    @Test
+    fun deleteDeviceToken_givenNoDeviceToken_expectDoNotAddTaskToBackgroundQueue() {
+        prefRepository.saveIdentifier(String.random)
+
+        repository.deleteDeviceToken()
+
+        verifyNoInteractions(backgroundQueueMock)
+    }
+
+    @Test
+    fun deleteDeviceToken_givenNoProfileIdentified_expectDoNotAddTaskToBackgroundQueue() {
+        prefRepository.saveDeviceToken(String.random)
+
+        repository.deleteDeviceToken()
+
+        verifyNoInteractions(backgroundQueueMock)
+    }
+
+    @Test
+    fun deleteDeviceToken_givenDeviceTokenAndIdentifiedProfile_expectAddTaskToBackgroundQueue() {
+        val givenDeviceToken = String.random
+        val givenIdentifier = String.random
+        prefRepository.saveDeviceToken(givenDeviceToken)
+        prefRepository.saveIdentifier(givenIdentifier)
+
+        repository.deleteDeviceToken()
+
+        verify(backgroundQueueMock).queueDeletePushToken(givenIdentifier, givenDeviceToken)
+    }
+}
